Turned LoginActivity into LoginDialog
