+++ android/app/src/main/res/layout/dialog_login.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
+<android.support.v7.widget.CardView
+        style="@style/Widget.MaterialComponents.CardView"
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".LoginActivity">
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:contentPadding="16dp">
 
     <RelativeLayout
             android:layout_width="200dp"
@@ -47,4 +48,4 @@
         </LinearLayout>
     </RelativeLayout>
 
-</FrameLayout>
+</android.support.v7.widget.CardView>
